+++ install_node_manager.sh
@@ -49,7 +49,7 @@ REPO_URL="${REPO_URL:-https://github.com/murat-taskaynatan/Blockdag-Node-Manager
 REPO_REF="${REPO_REF:-${REPO_BRANCH:-main}}"
 
 need_cmd "$PYTHON_BIN"
-ensure_packages python3 python3-venv python3-pip rsync
+ensure_packages python3 python3-venv python3-pip rsync fio
 
 TEMP_DIR=""
 cleanup() {